+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 
-- [Library] Elevation semenatic tokens `elevationBlurOverlayDefault`, `elevationXOverlayDefault`
+- [Library] Elevation semenatic tokens `elevationBlurOverlayDefault`, `elevationXOverlayDefault`, `elevationBlurOverlayEmphasized`
 
 ## [0.7.0](https://github.com/Orange-OpenSource/ouds-ios/compare/0.6.0...0.7.0) - 2024-11-29